Niti_Network/analytics/requirements.txt
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
joblib==1.4.0
